requirements.txt
asgiref==3.2.7
astroid==2.4.1
backcall==0.1.0
decorator==4.4.2
Django==3.0.4
django-crispy-forms==1.9.0
django-environ==0.4.5
django-extensions==2.2.9
django-taggit==1.2.0
gunicorn==20.0.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pylint==2.5.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.1.9
wrapt==1.12.1
